Added in Device Class from Develop
While This branch still isn't merged with develop, I took the device
class from develop and put it in here to make the addDevice method in
UtilController.

I might as well explain the changes that are WIP right here:

I added a method to UtilController called getPrinterInfo that creates a
device class with information on the printer (in this case, only column
names are taken from the DB). Theres a WIP class called submitBuildToDB
that takes in the build name, file name, and printer device. it will be
used when a person presses sumbit after they entered in their
information. It will store respective information in the right places in
the DB and change its status to completed.