Fixed the Student login and account creation stuff.
Made the student login and new student creation actually check against
existing ids in the database. also changed the schema a little bit to
add onecard columns to both users and admin tables, but not going to do
anything with those until I get feedback on how we wanna handle that.