Created methods to convert build process to dynamic
Added and adjusted many methods in UtilController to make previous
functions work in the dynamic database.

To replace the legacy method submitBuildInfoToDB is a new method called
submitBuildToDB. It serves the same purpose as the previous method, it
gathers the information entered in by the user on the Dialog page, and
sends it to the database. It also moves the printed file to the printed
directory of its respective printer and changes the status of any jobs
in the build to "completed".

To create this method, I had to create a few other methods in places
classes that are not a part of my task.

To the SQLMethods class, I added a duplicate of the
updatePendingJobsBuildName that takes in a String instead of an integer
because the schema has an varchar as the primary key for the build
entity and I needed to update the SQL command to match. I also added a
method called updateBuildData that is used by
UtilController.submitBuildToDB to take things the user has typed in and
store them in the DB, as this wasn't covered by the previous methods.

In FileMethods, I added two methods, getDevicePrinted and
getDeviceToPrint. These are used in the process of storing the submitted
files to the proper places on the OTS drive. This should cover the
change to a dynamic file system unless the system is going to
drastically change, and even then I don't think there would have to be
too much modifying to fix it.

Also, for temporary purposes, I made a PrinterDialogTemp class. I did
this to show how the submit method I made would relate to the eventual
Dialog class that's currently in the works.

Beside that method, there are a few other methods in UtilController I
created and modified to transition to dynamic:

getPrinterInfo creates a device class with the name of the printer and
the name of its custom columns. I thought this could be used to help
create the dynamic GUI, though it could be further expanded on if needed
elsewhere.

a makeBuild class that creates the initial build row in a DB. As such it
contains no information on the number of models or total runtime.

Everything should be set up for conversion to the dynamic database.